liblzma

Function encode_all

pub fn encode_all<R: Read>(source: R, level: u32) -> Result<Vec<u8>>

Compress from the given source as if using a read::XzEncoder.

The input data must be in the xz format. The level argument is typically 0-9 with 6 being a good default. To use the slower xz --extreme-style preset, bitwise-OR a level with stream::PRESET_EXTREME (for example, 6 | stream::PRESET_EXTREME).