liblzma

Module write

Writer-based compression/decompression streams

Structs

AutoFinishXzDecoder

A compression stream which will have compressed data written to it and will write uncompressed data to an output stream. AutoFinishXzDecoder impl Drop trait, so automatically calls XzDecoder::try_finish method when exiting the scope. However, it is not guaranteed that try_finish will complete successfully, and it is recommended to call try_finish manually if you want to ensure that the process is successful.

AutoFinishXzEncoder

A compression stream which will have uncompressed data written to it and will write compressed data to an output stream. AutoFinishXzEncoder impl Drop trait, so automatically calls XzEncoder::try_finish method when exiting the scope. However, it is not guaranteed that try_finish will complete successfully, and it is recommended to call try_finish manually if you want to ensure that the process is successful.

XzDecoder

A compression stream which will have compressed data written to it and will write uncompressed data to an output stream. XzDecoder will no longer perform the finalization automatically in the next miner release, so you need to call XzDecoder::finish manually. If you want to automate the finalization process, please use XzDecoder::auto_finish.

XzEncoder

A compression stream which will have uncompressed data written to it and will write compressed data to an output stream. XzEncoder will no longer perform the finalization automatically in the next miner release, so you need to call XzEncoder::finish manually. If you want to automate the finalization process, please use XzEncoder::auto_finish.